+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/connecting/PasswordLoginViewModel.kt
@@ -12,7 +12,6 @@ import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
-import net.folivo.trixnity.client.MatrixClient
 import org.koin.core.component.inject
 
 
@@ -22,7 +21,7 @@ interface PasswordLoginViewModelFactory {
     fun create(
         viewModelContext: ViewModelContext,
         serverUrl: String,
-        onLogin: (MatrixClient) -> Unit,
+        onLogin: () -> Unit,
         onBack: () -> Unit,
     ): PasswordLoginViewModel {
         return PasswordLoginViewModelImpl(
@@ -53,7 +52,7 @@ interface PasswordLoginViewModel {
 open class PasswordLoginViewModelImpl(
     viewModelContext: ViewModelContext,
     override val serverUrl: String,
-    private val onLogin: (MatrixClient) -> Unit,
+    private val onLogin: () -> Unit,
     private val onBack: () -> Unit,
 ) : ViewModelContext by viewModelContext, PasswordLoginViewModel {